backend/requirements.txt
# Core Data Processing
pandas~=2.2.3
numpy~=2.2.2
pyarrow<17.0.0,>=3.0.0

# Geospatial Processing
geopandas~=1.0.1
shapely~=2.0.6
pyproj~=3.7.0
fiona~=1.10.1

# Cloud Services
google-cloud-storage~=2.19.0
google-cloud-bigquery~=3.29.0
google-cloud-secret-manager~=2.22.1
fsspec~=2024.12.0
gcsfs~=2024.12.0

# Web Framework
fastapi~=0.115.7
uvicorn~=0.34.0
pydantic~=2.10.6

# HTTP and API Clients
aiohttp~=3.11.11
aiohttp[speedups]~=3.11.11
requests~=2.32.3
zeep~=4.3.1
paramiko~=3.5.0
certifi~=2024.12.14
pyOpenSSL~=25.0.0

# Data Pipeline and Validation
apache-beam~=2.62.0
backoff~=2.2.1

# PDF Processing
pdfplumber~=0.11.5

# Utilities
python-dotenv~=1.0.1
tqdm~=4.67.1
psutil~=6.1.1
cryptography~=44.0.0
lxml~=5.3.0
ijson~=3.3.0
